==> 1.No Escape from the Scrape. Instagram Scraping .ipynb <==
# Instagram Scraping 

# Data Analysis - Part I - Landscape of 50 most followed profiles

import pandas as pd
import numpy as np

most_followed = pd.read_csv('most_followed.csv', delimiter=';')

most_followed.head()

most_followed.tail()

## How many profiles where taken into consideration?

most_followed['id'].count()

## Which one has the most followers?

most = most_followed.sort_values(by=['number_of_followers'], ascending=False)

most.head()

# for readable numbers

most_followed['followers_in_millions'] = most_followed['number_of_followers'] / 1000000

most_followed.head()

most_readable = most_followed.sort_values(by=['followers_in_millions'], ascending=False)

most_readable.head()

most_readable.tail()

## Sum of all followers together

most_followed['number_of_followers'].sum()

most_readable['followers_in_millions'].sum()

# in billions:

most_readable['followers_in_millions'].sum() / 1000

## Which gender has most followers?

most_followed.groupby('gender').count()

most_followed.groupby('gender').sum()

## Which profession has the most followers?

most_followed.groupby('profession').sum().sort_values('number_of_followers', ascending=False)

# Part II - Analysis of posts from the most followed

post_details = pd.read_csv('post_details.csv', delimiter=';')

post_details.head()

## Which type of post receives the most likes?

post_details.info()

photo = post_details['post_type'] == 'photo'

post_details.info()

post_details['likes'] = post_details.likes.astype(str)

post_details.head()

post_details['likes'].apply(lambda x: x.replace(',',''))

post_details['likes'] = (post_details['likes'].replace('\.','', regex=True)
                        .astype(int))

post_details.head()

post_details.groupby('post_type').sum()

photo = post_details['post_type']=='photo'

post_details.where(photo).count()

average_on_photo = 3774093314 /1833

average_on_photo

video = post_details['post_type']=='video'

post_details.where(video).count()

average_on_video = 3547407445 /671

average_on_video

post_details.likes.min()

post_details.likes.max()

post_details.likes.mean()

post_details.likes.median()

post_details.where(post_details['post_type']=='photo').sort_values('likes', ascending=False).head()

## Best performing photo has Billie Eilish - https://www.instagram.com/p/B7JpPbgFOet/ - number 45 on the list of most popular profiles:

post_details.iloc[2161]

post_details.where(post_details['post_type']=='video').sort_values('likes', ascending=False).head()

## Best performing video belongs to Cristiano - https://www.instagram.com/p/B6RNPBsg1wo/ - He anyways has a profile with the most followers:

post_details.iloc[46]

post_details = pd.read_csv('post_details.csv', delimiter=';')

post_details['timezone'] = post_details['time_posted'][17:]

post_details.head()
