==> 2.Instanalysis_Data_Preparation.ipynb <==
# Preparing the dataset:
# Splitting the dataset into a training set and a set with new data

import pandas as pd
import numpy as np

post_comments_labelled = pd.read_csv('post_comments_labelled.csv', delimiter=';')

post_comments_labelled

post_comments_labelled.count()

post_comments_labelled = post_comments_labelled.drop(post_comments_labelled[post_comments_labelled.contents == 'Bestätigt'].index)

post_comments_labelled.to_csv('post_comments_labelled.csv', index=False)

post_comments_1 = post_comments_labelled.loc[post_comments_labelled['Sentiment']=='Positive']

post_comments_2 = post_comments_labelled.loc[post_comments_labelled['Sentiment']=='Negative']

post_comments_3 = post_comments_labelled.loc[post_comments_labelled['Sentiment']=='Neutral']

frames = [post_comments_1, post_comments_2, post_comments_3]

post_comments_train = pd.concat(frames)

post_comments_train.head()

post_comments_train.count()

shuffing the rows in the train dataframe:

post_comments_train_shuffled = post_comments_train.sample(frac=1).reset_index(drop=True)

turning emojis into their written names:

import emoji

post_comments_train_shuffled['contents'] = post_comments_train_shuffled['contents'].apply(emoji.demojize)

post_comments_train_shuffled.head()

saving the training dataframe into a csv file:

post_comments_train_shuffled.to_csv('post_comments_train.csv', index=False)

preparing the new unlabelled data:

post_comments_newdata = post_comments_labelled[post_comments_labelled.Sentiment.isna()]

shuffiling the newdata dataframe:

post_comments_newdata_shuffled = post_comments_newdata.sample(frac=1).reset_index(drop=True)

post_comments_newdata_shuffled['contents'] = post_comments_newdata_shuffled['contents'].apply(emoji.demojize)

from langdetect import DetectorFactory

from langdetect import detect

from langdetect.lang_detect_exception import LangDetectException

DetectorFactory.seed = 0

for x in post_comments_newdata_shuffled['contents']:
    try:
        if detect(x) != 'en':
            post_comments_newdata_shuffled.drop(post_comments_newdata_shuffled[post_comments_newdata_shuffled.contents == x].index)
    except LangDetectException:
        pass

and then saving the dataframe with new data into a csv file:

post_comments_newdata_shuffled.to_csv('post_comments_newdata.csv', index=False)

preparing a first batch of test data from the new data:

post_comments_test = post_comments_newdata_shuffled.loc[:100]

post_comments_test.to_csv('post_comments_test_NB.csv', index=False)
