==> 3.Instalysis_Part_checking_the_parameters.ipynb <==
# Checking which parameters work the best

import pandas as pd
from pprint import pprint
from sklearn.feature_extraction.text import CountVectorizer
from sklearn.feature_extraction.text import TfidfTransformer
from sklearn.linear_model import SGDClassifier
from sklearn.model_selection import GridSearchCV
from sklearn.pipeline import Pipeline

data = pd.read_csv('post_comments_train.csv', delimiter=',')

# data = data.Sentiment.map({'Negative':0, 'Neutral':1, 'Positive':2})
todrop = ['id', 'url', 'comment_author', 'time_posted', 'user_name_post_author']
data = data.drop(columns=todrop)

X = data.contents

Y = data.Sentiment

Y = Y.map({'Negative':0, 'Neutral':1, 'Positive':2})

pipeline = Pipeline([
    ('vect', CountVectorizer()),
    ('tfidf', TfidfTransformer()),
    ('clf', SGDClassifier()),
])

parameters = {
    'vect__max_df': (0.5, 0.75, 1.0),
    'vect__max_features': (None, 5000, 10000, 50000),
    'vect__ngram_range': ((1, 1), (1, 2)),  # unigrams or bigrams
    'tfidf__use_idf': (True, False),
    'tfidf__norm': ('l1', 'l2'),
    'clf__max_iter': (20,),
    'clf__alpha': (0.00001, 0.000001),
    'clf__penalty': ('l2', 'elasticnet'),
    'clf__max_iter': (10, 50, 100)
}

grid_search = GridSearchCV(pipeline, parameters, n_jobs=-1, verbose=1)

print("Performing grid search...")
print("pipeline:", [name for name, _ in pipeline.steps])
print("parameters:")
pprint(parameters)
grid_search.fit(X,Y)
print("Best score: %0.3f" % grid_search.best_score_)
print("Best parameters set:")
print(grid_search.best_estimator_)
best_parameters = grid_search.best_estimator_.get_params()
for param_name in sorted(parameters.keys()):
    print("\t%s: %r" % (param_name, best_parameters[param_name]))

# CountVectorizer() and SGDClassifier() had the best scores
