==> 4.Instalysis_LinearSVM.ipynb <==
# Analysis of the comments on Instagram

# Using ML to conduct sentiment analysis with LinearSVM

import pandas as pd
import numpy as np
import sklearn as sk

comments = pd.read_csv('post_comments_train.csv', delimiter=',')

comments.head()

todrop = ['id', 'url', 'comment_author', 'time_posted', 'user_name_post_author']
comments = comments.drop(columns=todrop)

comments.Sentiment = comments.Sentiment.map({'Negative':0, 'Neutral':1, 'Positive':2})

from sklearn.linear_model import SGDClassifier
from sklearn.pipeline import Pipeline
from sklearn.feature_extraction.text import CountVectorizer
from sklearn.feature_extraction.text import TfidfTransformer

text_clf_svm = Pipeline([('vect', CountVectorizer()),
                         ('tfidf', TfidfTransformer()),
                         ('clf-svm', SGDClassifier(loss='hinge', penalty='l2',alpha=1e-3, random_state=42))])

X_train = comments['contents']
Y_train = comments['Sentiment']

text_clf_svm.fit(X_train, Y_train)

# importing the test data:
test_data = pd.read_csv('post_comments_test.csv', delimiter=';')
test_data.Sentiment = test_data.Sentiment.map({'Negative':0, 'Neutral':1, 'Positive':2})

X_test = test_data['contents']
Y_test = test_data['Sentiment']

text_clf_svm.fit(X_train, Y_train)

predicted_svm = text_clf_svm.predict(X_test)

predicted_svm

np.mean(predicted_svm == Y_test)

from sklearn.model_selection import GridSearchCV

parameters_svm = {'vect__ngram_range': [(1, 1), (1, 2)],
                  'tfidf__use_idf': (True, False),
                  'clf-svm__alpha': (1e-2, 1e-3),}

gs_clf_svm = GridSearchCV(text_clf_svm, parameters_svm, n_jobs=-1)

gs_clf_svm = gs_clf_svm.fit(X_train, Y_train)

gs_clf_svm.best_score_

gs_clf_svm.best_params_

text_clf_svm = Pipeline([('vect', CountVectorizer(ngram_range=(1, 1))),
                         ('clf-svm', SGDClassifier(alpha=0.001, random_state=42))])

text_clf_svm.fit(X_train, Y_train)

predicted_svm = text_clf_svm.predict(X_test)

np.mean(predicted_svm == Y_test)
