==> 5. Instalysis_Naive_Bayes.ipynb <==
# Analysis of the comments on Instagram

# Using ML to conduct sentiment analysis with Naive Bayes

import pandas as pd
import numpy as np
import sklearn as sk

comments = pd.read_csv('post_comments_train.csv', delimiter=',')

comments.head()

every comment is a document - corpus <br>
we only need the column 'contents' and 'Sentiment', we will drop the other ones:

converting the Sentiment into a numerical variable:

comments.tail()

todrop = ['id', 'url', 'comment_author', 'time_posted', 'user_name_post_author']
comments = comments.drop(columns=todrop)

comments.Sentiment = comments.Sentiment.map({'Negative':0, 'Neutral':1, 'Positive':2})

creating a matrix from the database:

numpy_array = comments.to_numpy()

X = numpy_array[:,0]
Y = numpy_array[:,1]
Y = Y.astype(int)
print('X')
print(X)
print('Y')
print(Y)

creating word vectorization - bag od words:

from sklearn.feature_extraction.text import CountVectorizer 
vec = CountVectorizer()

vec.fit(X)
vec.vocabulary_

removing the stop words:

vec = CountVectorizer(stop_words='english' )
vec.fit(X)
vec.vocabulary_

print(vec.get_feature_names())
print(len(vec.get_feature_names()))

Every document needs to be converted into a feature vector representing presence of these words in that document. Let's convert each of our training documents in to a feature vector.

# another way of representing the features
X_transformed=vec.transform(X)
X_transformed

print(X_transformed[:5])

# converting transformed matrix back to a sparse matrix
X=X_transformed.toarray()
X

pd.DataFrame(X, columns=vec.get_feature_names())

# loading the testing data:
post_comments_testNB = pd.read_csv('post_comments_test_NB.csv', delimiter=';')

# converting label to a numerical variable
post_comments_testNB['Sentiment'] = post_comments_testNB.Sentiment.map({'Negative':0, 'Neutral':1, 'Positive':2})

dropcol = ['id', 'url', 'comment_author', 'time_posted', 'user_name_post_author']
post_comments_testNB = post_comments_testNB.drop(columns=dropcol)

test_matrix = post_comments_testNB.to_numpy()
X_testNB = test_matrix[:,0]
Y_testNB = test_matrix[:,1]
Y_testNB = Y_testNB.astype('int')
print("X_test")
print(X_testNB)
print("Y_test")
print(Y_testNB)

X_testNB_transformed=vec.transform(X_testNB)

X_testNB=X_testNB_transformed.toarray()
X_testNB

# Multinomial Naive Bayes¶

The first supervised learning method we introduce is the multinomial Naive Bayes or multinomial NB model, a probabilistic learning method. The proba- bility of a document d being in class c is computed as\
\
P(c|d) ∝ P(c) ∏ P(tk|c) 1≤k≤nd\
                    \
where P(tk|c) is the conditional probability of term tk occurring in a docu- ment of class c.1 We interpret P(tk|c) as a measure of how much evidence tk contributes that c is the correct class. P(c) is the prior probability of a document occurring in class c. If a document’s terms do not provide clear evidence for one class versus another, we choose the one that has a higher prior probability. ⟨t1, t2, . . . , tnd ⟩ are the tokens in d that are part of the vocab- ulary we use for classification and nd is the number of such tokens in d. For example, ⟨t1, t2, . . . , tnd ⟩ for the one-sentence document Beijing and Taipei join the WTO might be ⟨Beijing, Taipei, join, WTO⟩, with nd = 4, if we treat the terms and and the as stop words.

# building a multinomial NB model
from sklearn.naive_bayes import MultinomialNB

# instantiate NB class
mnb=MultinomialNB()

# fitting the model on training data
mnb.fit(X,Y)

# predicting probabilities of test data
mnb.predict_proba(X_testNB)


proba=mnb.predict_proba(X_testNB)
pd.DataFrame(proba, columns=['Neutral','Negative', 'Positive'])

mnb.score(X,Y)

# Bernoulli Naive Bayes\
\
An alternative to the multinomial model is the multivariate Bernoulli model or Bernoulli model. It generates an indicator for each term of the vo- cabulary, either 1 indicating presence of the term in the document or 0 indi- cating absence. The Bernoulli model has the same time complexity as the multinomial model.
The different generation models imply different estimation strategies and different classification rules. The Bernoulli model estimates Pˆ(t|c) as the fraction of documents of class c that contain term t

from sklearn.naive_bayes import BernoulliNB

# instantiating bernoulli NB class
bnb=BernoulliNB()

# fitting the model
bnb.fit(X,Y)

# predicting probability of test data
bnb.predict_proba(X_testNB)
proba_bnb=bnb.predict_proba(X_testNB)

pd.DataFrame(proba_bnb, columns=['Neutral','Negative','Positive'])

bnb.score(X,Y)
